EventDto moved to separate file. Added unsubscription
