+++ libs/event-log/src/lib/ui-event-list/event-list.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
+import { Component, OnInit, Input, Output, EventEmitter, OnDestroy } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { AppState } from '../+state/index';
 import { changePage } from '../+state/actions/filters.action';
